requirements.txt
streamlit
opencv-python
scikit-image
imageio
Pillow 